API

IntervalSets.AbstractInterval — Type

A subtype of AbstractInterval{T} represents an interval subset of type T, that provides endpoints, closedendpoints.

IntervalSets.ClosedInterval — Type

A ClosedInterval(left, right) is an interval set that includes both its upper and lower bounds. In mathematical notation, the constructed range is [left, right].

IntervalSets.Domain — Type

A subtype of Domain{T} represents a subset of type T, that provides in.

IntervalSets.Interval — Type

An Interval{L,R}(left, right) where L,R are :open or :closed is an interval set containg x such that

left ≤ x ≤ right if L == R == :closed

left < x ≤ right if L == :open and R == :closed

left ≤ x < right if L == :closed and R == :open, or

left < x < right if L == R == :open

IntervalSets.OpenInterval — Type

An TypedEndpointsInterval{:open,:open}(left, right) is an interval set that includes both its upper and lower bounds. In mathematical notation, the constructed range is (left, right).

IntervalSets.TypedEndpointsInterval — Type

A subtype of TypedEndpointsInterval{L,R,T} where L and R are :open or :closed, that represents an interval subset of type T, and provides endpoints.

Base.Math.clamp — Method

clamp(t, i::ClosedInterval)

Clamp the scalar t such that the result is in the interval i.

Base.findall — Method

findall(in(interval), x::AbstractRange)

Return all indices i for which x[i] ∈ interval, specialized for the case where x is a range, which enables constant-time complexity.

Examples

julia> x = range(0,stop=3,length=10)
0.0:0.3333333333333333:3.0

julia> collect(x)'
1×10 adjoint(::Vector{Float64}) with eltype Float64:
 0.0  0.333333  0.666667  1.0  1.33333  1.66667  2.0  2.33333  2.66667  3.0

julia> findall(in(1..6), x)
4:10

It also works for decreasing ranges:

julia> y = 8:-0.5:0
8.0:-0.5:0.0

julia> collect(y)'
1×17 adjoint(::Vector{Float64}) with eltype Float64:
 8.0  7.5  7.0  6.5  6.0  5.5  5.0  4.5  …  3.0  2.5  2.0  1.5  1.0  0.5  0.0

julia> findall(in(1..6), y)
5:15

julia> findall(in(Interval{:open,:closed}(1,6)), y) # (1,6], does not include 1
5:14

Base.mod — Method

mod(x, i::AbstractInterval)

Find y in the i interval such that $x ≡ y (mod w)$, where w = width(i).

Examples

julia> I = 2.5..4.5;

julia> mod(3.0, I)
3.0

julia> mod(5.0, I)
3.0

julia> mod(2.5, I)
2.5

julia> mod(4.5, I)  # (a in I) does not imply (a == mod(a, I)) for closed intervals
2.5

julia> mod(4.5, Interval{:open, :closed}(2.5, 4.5))
4.5

Base.range — Method

range(i::ClosedInterval; step, length)
range(i::ClosedInterval, len::Integer)

Constructs a range of a specified step or length.

Base.range — Method

range(i::Interval{:closed,:open}; length)
range(i::Interval{:closed,:open}, len::Integer)

Constructs a range of a specified length with step=width(i)/length.

IntervalSets.:.. — Method

iv = l..r

Construct a ClosedInterval iv spanning the region from l to r.

IntervalSets.:± — Method

iv = center±halfwidth

Construct a ClosedInterval iv spanning the region from center - halfwidth to center + halfwidth.

IntervalSets.closedendpoints — Method

A tuple of Bool's encoding whether the left/right endpoints are closed.

IntervalSets.endpoints — Method

A tuple containing the left and right endpoints of the interval.

IntervalSets.isleftclosed — Method

Is the interval closed at the left endpoint?

IntervalSets.isleftopen — Method

Is the interval open at the left endpoint?

IntervalSets.isopenset — Method

Is the interval open?

IntervalSets.isrightclosed — Method

Is the interval closed at the right endpoint?

IntervalSets.isrightopen — Method

Is the interval open at the right endpoint?

IntervalSets.leftendpoint — Method

The left endpoint of the interval.

IntervalSets.rightendpoint — Method

The right endpoint of the interval.

IntervalSets.searchsorted_interval — Method

searchsorted_interval(a, i::Interval; [rev=false])

Return the range of indices of a which is inside of the interval i (using binary search), assuming that a is already sorted. Return an empty range located at the insertion point if a does not contain values in i.

Examples

julia> searchsorted_interval([1,2,3,5], 2..4)
2:3

julia> searchsorted_interval([1,2,3,5], 4..1)
4:3

julia> searchsorted_interval(Float64[], 1..3)
1:0

IntervalSets.width — Method

w = width(iv)

Calculate the width (max-min) of interval iv. Note that for integers l and r, width(l..r) = length(l:r) - 1.